+++ scm-ui/ui-components/src/repos/index.ts
@@ -19,6 +19,7 @@ export * from "./changesets";
 
 export { default as Diff } from "./Diff";
 export { default as DiffFile } from "./DiffFile";
+export { default as DiffButton } from "./DiffButton";
 export { default as LoadingDiff } from "./LoadingDiff";
 export { DefaultCollapsed, DefaultCollapsedFunction } from "./defaultCollapsed";